+++ Todo/Extension/UIColor.swift
@@ -1,0 +1,13 @@
+//
+//  ExtentsionUIColor.swift
+//  Todo
+//
+//  Created by 김도현 on 2023/08/24.
+//
+
+import Foundation
+import UIKit
+
+extension UIColor {
+    static var mainColor: UIColor = UIColor(red: 109/255, green: 209/255, blue: 213/255, alpha: 1.0)
+}
